refactor: [NESTJS] Rename users/entity to users/entities
