+++ packages/api-nestjs/src/auth/dto/auth-response.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { User } from '../../users/entity/user.entity';
+import { User } from '../../users/entities/user.entity';
 
 export class AuthUserResponse {
     @ApiProperty({ type: String, nullable: true })
